chore: 큰 병 컴포넌트 내 폴더 구조 변경
