+++ apps/client/src/widgets/report/components/keunbyeong/components/noehyeolgwan.tsx
@@ -1,40 +1,44 @@
 import { Alert } from '@bds/ui';
 
-import { Accordion } from '../accordion/accordion';
-import Graph from '../graph/graph';
-import Info from '../info/info';
-import Title from '../title/title';
-import { shimjangData } from './mocks/keunbyeong-mocks';
+import { Accordion } from '../../accordion/accordion';
+import Graph from '../../graph/graph';
+import Info from '../../info/info';
+import Title from '../../title/title';
+import { noehyeolgwanData } from '../mocks/keunbyeong-mocks';
 
 import * as styles from './style.css';
 
 const SECTION = [
-  { title: '급성 심근경색', key: 'acuteMyocardialInfarction' },
-  { title: '허혈성 심장질환', key: 'ischemic' },
-  { title: '확대 심장질환', key: 'extended' },
-  { title: '부정맥, 심부전', key: 'arrhythmia' },
+  { title: '뇌출혈', key: 'hemorrhage' },
+  { title: '뇌졸증', key: 'infarction' },
+  { title: '기타 뇌혈관질환', key: 'other' },
 ] as const;
 
-const Shimjang = () => {
+const Noehyeolgwan = () => {
   return (
     <Accordion>
-      <Accordion.Header type="강력"> 심장질환</Accordion.Header>
+      <Accordion.Header type={noehyeolgwanData.coverageStatus}>
+        뇌혈관질환
+      </Accordion.Header>
       <Accordion.Panel>
         <Info
-          description={shimjangData.additional_info}
+          description={noehyeolgwanData.additional_info}
           size="sm"
           iconSize="1.6rem"
         />
         <div className={styles.allgraphContainer}>
           {SECTION.map(({ title, key }) => {
-            const diagnosisCoverage =
-              shimjangData[key].diagnosis.productCoverage;
-            const injuryCoverage = shimjangData[key].injury.productCoverage;
+            const diagnosisCoverage = noehyeolgwanData[key].diagnosis
+              .productCoverage as number;
+            const injuryCoverage = noehyeolgwanData[key].injury
+              .productCoverage as number;
             const showAllAlert =
               diagnosisCoverage === 0 && injuryCoverage === 0;
+
             return (
               <div key={key} className={styles.graphContainer}>
                 <Title category="subCategory" title={title} />
+
                 <div className={styles.graphContentsContainer}>
                   {showAllAlert ? (
                     <Alert
@@ -50,7 +54,9 @@ const Shimjang = () => {
                       {diagnosisCoverage !== 0 ? (
                         <Graph
                           detailItem="진단비"
-                          average={shimjangData[key].diagnosis.averageCoverage}
+                          average={
+                            noehyeolgwanData[key].diagnosis.averageCoverage
+                          }
                           current={diagnosisCoverage}
                         />
                       ) : (
@@ -61,15 +67,14 @@ const Shimjang = () => {
                             iconSize="2rem"
                             alertHeader="참고하세요!"
                             alertContents="는 이 보험에 포함되지 않아요."
-                            highlight={`진단비`}
+                            highlight="진단비"
                           />
                         </div>
                       )}
-
                       {injuryCoverage !== 0 ? (
                         <Graph
                           detailItem="수술비"
-                          average={shimjangData[key].injury.averageCoverage}
+                          average={noehyeolgwanData[key].injury.averageCoverage}
                           current={injuryCoverage}
                         />
                       ) : (
@@ -80,7 +85,7 @@ const Shimjang = () => {
                             iconSize="2rem"
                             alertHeader="참고하세요!"
                             alertContents="는 이 보험에 포함되지 않아요."
-                            highlight={`수술비`}
+                            highlight="수술비"
                           />
                         </div>
                       )}
@@ -96,4 +101,4 @@ const Shimjang = () => {
   );
 };
 
-export default Shimjang;
+export default Noehyeolgwan;
