+++ apps/client/src/widgets/report/components/keunbyeong/components/shimjang.tsx
@@ -1,38 +1,37 @@
 import { Alert } from '@bds/ui';
 
-import { Accordion } from '../accordion/accordion';
-import Graph from '../graph/graph';
-import Info from '../info/info';
-import Title from '../title/title';
-import { cancerData } from './mocks/keunbyeong-mocks';
+import { Accordion } from '../../accordion/accordion';
+import Graph from '../../graph/graph';
+import Info from '../../info/info';
+import Title from '../../title/title';
+import { shimjangData } from '../mocks/keunbyeong-mocks';
 
 import * as styles from './style.css';
 
 const SECTION = [
-  { title: '일반암', key: 'general' },
-  { title: '소액암', key: 'atypical' },
+  { title: '급성 심근경색', key: 'acuteMyocardialInfarction' },
+  { title: '허혈성 심장질환', key: 'ischemic' },
+  { title: '확대 심장질환', key: 'extended' },
+  { title: '부정맥, 심부전', key: 'arrhythmia' },
 ] as const;
 
-const Cancer = () => {
+const Shimjang = () => {
   return (
     <Accordion>
-      <Accordion.Header type={cancerData.coverageStatus}>암</Accordion.Header>
+      <Accordion.Header type="강력"> 심장질환</Accordion.Header>
       <Accordion.Panel>
         <Info
-          description={cancerData.additional_info}
+          description={shimjangData.additional_info}
           size="sm"
           iconSize="1.6rem"
         />
         <div className={styles.allgraphContainer}>
           {SECTION.map(({ title, key }) => {
-            const diagnosisCoverage = cancerData[key].diagnosis
-              .productCoverage as number;
-            const injuryCoverage = cancerData[key].injury
-              .productCoverage as number;
-
+            const diagnosisCoverage =
+              shimjangData[key].diagnosis.productCoverage;
+            const injuryCoverage = shimjangData[key].injury.productCoverage;
             const showAllAlert =
               diagnosisCoverage === 0 && injuryCoverage === 0;
-
             return (
               <div key={key} className={styles.graphContainer}>
                 <Title category="subCategory" title={title} />
@@ -51,8 +50,8 @@ const Cancer = () => {
                       {diagnosisCoverage !== 0 ? (
                         <Graph
                           detailItem="진단비"
-                          average={cancerData[key].diagnosis.averageCoverage}
-                          current={cancerData[key].diagnosis.productCoverage}
+                          average={shimjangData[key].diagnosis.averageCoverage}
+                          current={diagnosisCoverage}
                         />
                       ) : (
                         <div className={styles.alertDiagnosisContainer}>
@@ -62,15 +61,16 @@ const Cancer = () => {
                             iconSize="2rem"
                             alertHeader="참고하세요!"
                             alertContents="는 이 보험에 포함되지 않아요."
-                            highlight="진단비"
+                            highlight={`진단비`}
                           />
                         </div>
                       )}
+
                       {injuryCoverage !== 0 ? (
                         <Graph
                           detailItem="수술비"
-                          average={cancerData[key].injury.averageCoverage}
-                          current={cancerData[key].injury.productCoverage}
+                          average={shimjangData[key].injury.averageCoverage}
+                          current={injuryCoverage}
                         />
                       ) : (
                         <div className={styles.alertInjuryContainer}>
@@ -80,7 +80,7 @@ const Cancer = () => {
                             iconSize="2rem"
                             alertHeader="참고하세요!"
                             alertContents="는 이 보험에 포함되지 않아요."
-                            highlight="수술비"
+                            highlight={`수술비`}
                           />
                         </div>
                       )}
@@ -96,4 +96,4 @@ const Cancer = () => {
   );
 };
 
-export default Cancer;
+export default Shimjang;
